Updated APRS protocols and much more - Removed secondary decoder - Updated image encoding at tracker firmware - Updated image encoding at decoder software - Implemented more sophisticated website - Added error transmission - Adjusted MCU speed to the highest rate - Implemented configuration that allows GPS to be switched on permanently at a certain battery voltage - Decreased humidity sensor accuray - Imcreased voltage measurement accuracy - Added temperature measurements of STM32 and Si4464 to TrackPoint struct - Added light intensity measurements of OV5640 to TrackPoint struct - Cleaned up (removed unused things) - Implemented more accurate measurements for PAC1720 - Implemented failure detection for PAC1720 and ublox chips - Fixed bug in Si4464 driver which used power when switched off - Added more data readouts for GPS receiver (pDOP, gpsFixOK) - Changed way of determination if GPS fix is good - Implemented temperature measurement of STM32 - Added error detection of I2C bus - Removed Milliseconds from ptime_t - Fixed bug in timestamp conversion - Removed RBAT measurement implementation (it was very inaccurate) - Avoid radio being switched off when next packet is waiting for being transmitted - Tidied up tracking manager

+++ tracker/software/drivers/bme280.h
@@ -90,7 +90,7 @@ bool BME280_isAvailable(uint8_t address);
 void BME280_Init(bme280_t *handle, uint8_t address);
 int16_t BME280_getTemperature(bme280_t *handle);
 uint32_t BME280_getPressure(bme280_t *handle, uint16_t means);
-uint16_t BME280_getHumidity(bme280_t *handle);
+uint8_t BME280_getHumidity(bme280_t *handle);
 int32_t BME280_getAltitude(uint32_t seaLevel, uint32_t atmospheric);
 
 #endif
